testing: enforce that .pyi files are always tested
Being able to run .pyi files in Python 3 is always assumed
possible. Since this test is disabled in local builds, we
were relying on correct configuration of COIN, but that failed.

To be safe, we now add a test to wheel_tester as well that
unconditionally tests all .pyi with all features enabled.

A special problem was a name clash in Qt3DAnimation.pyi
which uses the name "property" :-)

An import of QtMultimedia seems not to work, always.
This happened in wheel_tester.py in CI, only.

Task-number: PYSIDE-1599
Change-Id: Ib158e710cec72287fe4a71c01254727ea9b6dc54
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 4269e3535b7b97c188a6d8377e685586da89592f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>