translate: Fix translation for derived types
[ChangeLog][PySide6] The `tr` function now works also for
PySide derived Python class instances with the right context.

Qt has a translation feature that works fine on strings that
are directly parsed from a Python class.

For derived classes, the correct context could not be found.

This problem could have been solved by obtaining the calling
function name from the frame stack, walking the mro and
examining in which class dict the calling function can be
found, but this works for Python functions, only.

For any function, simply probing the translation result is
even easier and always works.
We walk the mro, too, but simply use each class name in turn
as context and probe by comparison if a translation was done.

Change-Id: Ibca9dc0098a3aa26a674b3d5e24ea14e651fe6de
Fixes: PYSIDE-131
Pick-to: 6.1
Pick-to: 5.15
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>