qml: replace context properties and code updates
Most of the qml code in the repository was outdated,
and followed bad practices, like context properties.
Complementary, after the major updates for Qt6
most of the code was not relying on the new ways of
register types (singletons, and using the decorator QmlElement).

Drop the context property usage in the following examples:
- signals/qmltopy1
- signals/qmltopy2
- signals/pytoqml2
- usingmodel
- quickcontrols2/gallery
- textproperties

Additionally:
- all the tests related to context properties
- tutorials/qmlapp
- tutorials/qmlsqlintegration
- Removing 'scrolling' example
- Fixing some flake8 warnings

Change-Id: I649248c0149876bf2bf94e78e27cef7110f42f1d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Keith Kyzivat <keith.kyzivat@qt.io>