PySide6: Fix crashes in QObject.findChild()
For all objects encountered in findChild(), a wrapper
was created just to be able to retrieve its type object
and check against it. As there is a name lookup involved,
it happens that non-QObject wrappers are created for this
(in the specific case, an enum wrapper), which leads to
crashes later on.

Refactor the code, extracting a helper function to PySide
which finds the best-matching type object for a QObject
to check on that. Rearrange the check so that the name is
checked first.

Fixes: PYSIDE-1609
Change-Id: I026854201f3f6eca9e5905105127f0a4b4588fa9
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit 98aef951920641dcdf4c217a0c51626cdf82caf8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>