Pyside6/Qt Designer: Fix Python code preview not working on UNIX
Qt Designer as bundled by PySide6 was unable to find the uic
binary in the libexec directory of the bundled Qt since that was only
copied when QtWebEngine was built and the rcc/uic binaries were copied
into the main directory.

Also, libexec existed as a file containing qt.conf, which was created
by a copy statement not checking for the target directory.

Fix that by actually creating a libexec directory for  uic, rcc and
QtWebEngineProcess. Patch the executables accordingly.
Add checks before copying qt.conf.
Adapt pyside-tool to use libexec.

The Windows code path remains the same, everything uses
main directory there.

Change-Id: I0c9f46fb776ed0315eecb6ed00202ea8d8f17fe2
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Tismer <tismer@stackless.com>