Add fallback case for QSettings.value
This includes a case where the passed type is not
a Python type, so we fallback to a QVariant type
as Qt/C++ does.

A 'TypeError' is raised if the Python type used
is not considered.

Added a couple of tests too.

Fixes: PYSIDE-1619
Change-Id: I5f2e277bb6ccbf9c3a1b23e79717c11ff7635ce5
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit 07b93a82afc22760f87a68fa9926dd316c4a0815)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>