PySide: fix QItemSelection default and QItemSelection.__add__
QItemSelection with no argument should create an empty
selector. This worked in Qt5.

In Qt6, everything works but the empty argument.
Without further analysis, this might be related to the
new [default] tag found in the Qt6 QItemSelection documentation?

Fixing this bug leads directly to the add operator, which returns
a list instead of a combined QItemSelection.

Fixed by removing `operator+(list<QItemSelectionRange>)` and re-adding it
with QItemSelection

Unrelated, occurred during bug hunting.

Task-number: PYSIDE-535
Change-Id: Ie5b881659e54fc0eebc8c9903df6e14eb2788565
Pick-to: 6.1
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>