pathlib: fix empty constructors in conversion
In some classes, like QDir, there is a default value
for constructors that accept a QString:

    QDir(const QString &path = QString())

However this case was not considered in Python land
when nothing is passed.
Considering other classes might have a similar situation,
we fix this at the level of the conversion and not per Qt class.

To reproduce this issue, a simple:

    from PySide6.QtCore import QDir
    a = QDir()

will segfault.

Added a simple test case,
and instructions to make the parameter optional (pyi).

Task-number: PYSIDE-1499
Change-Id: I42156b87ca0aa60466c743a8cc8b42ea5eeb3559
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>