shiboken6: Add opaque containers for C++ sequence containers
Add a class that directly wraps a C++ sequence container,
allow for modifying them.

For all instantiated containers, generate a special (sequence) type
that wraps the C++ container directly. For example, it will be
accessible as a QList_int.

This is achieved via providing a template for a type private
that relies on a conversion traits template for conversion.
Only the  conversion traits specialization code needs to be generated.

Use cases:
- Allowing for modifying Fields of such container types
  (non-owning)
- Pass it into functions taking such containers instead of converting
  back and forth from a PyList (constructed in Python, owning)

[ChangeLog][shiboken6] Support for opaque C++ sequence scontainers
has been added, allowing to pass a wrapped C++ container
directly instead of converting it back and forth from
Python sequences.

Task-number: PYSIDE-1605
Change-Id: I49d378eb1a0151730d817d5bdd4b71a7c3b5cdda
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>