signature: make zip file access totally virtual
With the new implementation of an importer for virtual
zipfiles, there is no longer a problem with traces of files
in the file system.

Especially, cx_freeze should have no longer any signature
related problem.

This version cannot be backported to Python 2.7, but it will
work for Python 3 and PySide 5.15 .

[ChangeLog][shiboken6] Embedding of supporting Python files
is now completely virtual. No FS files are involved any longer.

Change-Id: Ifa0942b4476bff95e823505897b867735418ca69
Pick-to: 5.15
Pick-to: 6.1
Fixes: PYSIDE-1621
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>