feature: fix the UIC switching problem
The BindingManager::getOverride function computes the current
switch state from information of a type object. But the type object
must first be updated in case a switch has happened.

The solution was an extra update call at the beginning of the
function.

This solution _always_ works, with or without inheritance, for
Python >= 3.7.

[ChangeLog][shiboken6] Coexistence of different feature
selections works now, especially for UIC files and inheritance.

Fixes: PYSIDE-1626
Pick-to: 6.1
Pick-to: 5.15
Change-Id: I577331cfb2d7511110d1e16e729bed80985340a0
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>