Remove signature loader completely from physical FS
This proposal was discussed.
It was also tested with Nuitka and cx_freeze.

[ChangeLog][shiboken6] Python support files are now always
embedded. There is no longer a physical copy installed in the FS.

Reasoning is: We always had the signature module in real
Python files and also as an embedded ZIP file.

With the solution to PYSIDE-1621 we suddenly had a way
to avoid accessing the file system completely. The
remaining question is: Why should we keep the files in
the installation at all?

Long time ago, it was said that the source files were
needed to see tracebacks. But tracebacks are visible the
same way, regardless if the files exist as physical files
or only in memory. The error messages will be as verbose
as always.

Personally, I see it more as a drawback, because parts of
our application become vulnerable by external manipulation.
By the complete virtualization, the system is as
self-contained as a binary executable.

Task-number: PYSIDE-1621
Change-Id: I821545fb5d52465b69bb2c172bdfb43894ac2109
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>