shiboken6: Introduce a new normalization function for added functions
QMetaObject::normalizedSignature() was used for normalizing
added function signatures as well as matching modify-function
signatures. The function replaces const-ref by value which
is convenient for matching existing functions, but very problematic
for adding functions. It also does not work for added functions
with parameter names ("const QVariant &@name@" will be left unchanged).

Introduce a new function that only normalizes white space for
the add function cases.

[ChangeLog][shiboken] A new, simpler normalization function has
been introduced for signatures of added functions (which means
for example that const-ref is preserved).

Change-Id: I642aa484095523e3dabb263fadedb20e3c0d3f15
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Tismer <tismer@stackless.com>