Shiboken: Remove cheating macros
There are a number of cheating macros that allow to use
the same code in Python 2 and 3. Because Python 2 is gone,
remove these macros.

This conversion was partially difficult since certain types
collapsed in the XML files and generated functions contained
substrings of the macros. This is actually the fourth attempt.

Task-number: PYSIDE-1019
Pick-to: 6.1
Change-Id: I116877afc8aa36f4710a40df1769f600b6b750ea
Reviewed-by: Christian Tismer <tismer@stackless.com>