py3.10-prep: Fix a very old refcounting error in time_test.py
Since 2012-03-08, there was a refcounting error in
the typesystem_sample.xml file at <value-type name="Time">.
The boolean function results were not refcounted.

(Commit 7b731d702c64d172e192aa1876e4a6fade6c38d7)

This bug was hard to find because nobody expects a failure in the
testing code, and the error shows up at the very end in the
finalization code, totally unrelated to the error location.

In order to find it, the setting `PYTHONMALLOC=malloc` was needed,
because otherwise we got a Python memory dump but no traceback.
Then, Python was modified and code was inserted in the various
involved `code_dealloc`, `tupledealloc` and `object_dealloc`
functions that printed the involved refcounts and types before
performing the actual deallocation.

This bug actually was of no harm since only the wrong test was
involved. It was very misleading because there were quite some
changes to the `datetime` module in Python 3.10 . But the bug
was nothing but a dormant refcounting error.

[ChangeLog][shiboken6] A very old refcounting bug from 2012 in
the sample XML lead to a crash in Python 3.10 . This bug was
found because the refcounting of Python is more exact, now.

Task-number: PYSIDE-1436
Change-Id: I4363aae48e23f8ccdbd85c30654a44edbd69e516
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>