py3.10-prep: Fix a binary operator bug in cppgenerator.py
The cppgenerator produced on binary flag operators only checks
for cppArg but not for cppSelf. This is wrong, because arguments
are sometimes swapped on operator calls.

This was recognized in a debug Python build on Python 3.10 .
The error checking of this version has a lot improved.

[ChangeLog][shiboken6] Binary flags operators did check  cppArg
only, but not cppSelf. This is necessary when arguments are swapped.
Recognized on Python 3.10 due to improved error checks in debug mode.

Task-number: PYSIDE-1436
Change-Id: If14e295e6a5a55a6b648c711b7a934733858636a
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 7034d2ff83248cf9f1a959fdeabf47e5d20c8df8)
Reviewed-by: Christian Tismer <tismer@stackless.com>