Fix the signature mapping after enabling macOS primitive types
After the new primitive types appeased in macOS, there are
many new missing signatures. Many missing array modifications
could be solved im mapping.py .

The rest is only possible with a heavy change to the QtGui
and QtOpenGL XML files. It will be fixed in 6.2 when we have
the possibility to unify types by an XML PI.

Pick-to: 6.1
Task-number: PYSIDE-1646
Change-Id: I6e4c1b4e2409f830523d308b2e6dfafa3870132b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>