shiboken6: Split out node classes from OverloadData
OverloadData represents a tree, the root node of which has
the list of all functions and a list of children, which represent
an argument/type combination.

This was all in one class, where it was unclear which member function
or member variable was applicable to the node type.

To fix this, split out a root node class and an argument node class.

Rename nextOverloadData() to children() and previousOverloadData()
to parent() to make the relationship clearer.

Task-number: PYSIDE-1653
Change-Id: Ife63e78604bc42c08cc8cd0056d1c810060ba886
Reviewed-by: Christian Tismer <tismer@stackless.com>