shiboken6: Refactor writeMinimalConstructorExpression()
Change ShibokenGenerator::minimalConstructorExpression()
to return a string and remove the defaultCtor
parameter, which is only relevant for the argument conversion
case where the argument has a default value.

Put the default value code into a separate helper
in CppGenerator and make it a bit smarter, trying
to avoid assignments for class types.

Task-number: PYSIDE-1605
Change-Id: I22594cedcf8710fc85ae255109ab4ead1effcfa1
Reviewed-by: Christian Tismer <tismer@stackless.com>