PyPySide: Encapsulate type generation to be PyPy compatible
Type generation in PyPy is critical. When type creation calls
PyType_Ready, PyPy freezes the current state of the type object.
That has fatal effects, because with the Limited API, we need
to patch some types a little _after_ creation.

The functionality of this patch is still the same, but type
creation is totally encapsulated, and we can do a different
implementation for PyPy.

[ChangeLog][PySide6] PyPySide: Type creation was completely
encapsulated. The implementation can now be rewritten for PyPy.

Task-number: PYSIDE-535
Change-Id: Iae0cadb15b631344ae76cea8c104d4b69941c2b0
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>