shiboken6: Rewrite conversions
Currently, the isPythonToCpp(Value/Reference/Pointer)Convertible
functions used by the overload decisor return a function pointer. In the case
of isPythonToCppReferenceConvertible, this can be either a pointer
conversion taking a Foo** or a value conversion taking a Foo*.

The function isImplicitConversion(SbkObjectType *, PythonToCppFunc)
is called further down in the generated code to check whether it
is a value or a pointer.

In order to get rid of this, introduce a new version of the
checking function that returns a struct that has the function
and knows its type.
Add an overload of isPythonToCppReferenceConvertible() taking
a SbkConverter for the upcoming opaque containers.

Task-number: PYSIDE-1605
Change-Id: I49f70b85d01500d95208f0f2dd290bf23cb80eff
Reviewed-by: Christian Tismer <tismer@stackless.com>