Fix build for non-limited API
e55c87131a66e1cd11b13663b66582412cd38788 broke the build
for the non-limited API due to removing a C-style cast from
a code snippet:

%0 = PyBytes_GET_SIZE((PyObject *)%PYARG_0);

in the overridden method of a C++ wrapper class.
PyBytes_GET_SIZE(op) is defined to use PyVarObject*)(op)),
which does not work when passing an AutoDecRef instance
which is used for pyResult in that case.

To fix this, make the placeholder a smarter; append
.object() unless the code snippet invokes a member
function itself.

[ChangeLog][shiboken6] The replacement of %PYARG_0 in
C++ wrapper classes (native) has been improved and no
longer requires adding casts to PyObject *.

Change-Id: Ia31b28d568603f96c93e564e5ef7554cb2468b04
Reviewed-by: Christian Tismer <tismer@stackless.com>