shiboken6: Refactor modified types of functions
Remove AbstractMetaFunction::typeReplaced(), which returned the
modified type of an argument or the return type as a string.

Instead, for function arguments, for which modified types must always
be valid AbstractMetaTypes, add a getter modifiedType() returning an
AbstractMetaType to AbstractMetaArgument.

This fixes a number of potential bugs when accessing typeReplaced(int)
with the wrong position due to removed arguments.

The modified argument types are parsed from the XML strings in early on
in AbstractMetaClass::fixFunctions() with proper error reporting.

For return types, for which modified types need to be valid only when
generating a type check for virtual method return in the C++ wrapper,
add a function modifiedTypeName() to AbstractMetaFunction.

Adapt the client code accordingly.

Split apart CppGenerator::writeTypeCheck() into one overload
taking an AbstractMetaType and one taking a string.

Task-number: PYSIDE-1660
Change-Id: Id95379892b40ae2632ca33ed27263bada0ec015d
Reviewed-by: Christian Tismer <tismer@stackless.com>