shiboken6: Simplify CppGenerator::getArgumentType()
The function took an argument position where a value of 0 means return
type. The return type code branch was not used; it is only ever
invoked for arguments.

Remove the return type code path and change the function to take the
argument index. Failing to build a meta type for an argument is a hard
failure; change the function to return the meta type and to throw on
failure.

Task-number: PYSIDE-1660
Change-Id: I0dd89ecde595c71b42df7464d03204bafe74d17c
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>