shiboken6: Remove ShibokenGenerator::guessCPythonCheckFunction()
Since all CPython types are now built-in custom types, there is no
longer a need to guess the check functions.

Custom types passed to writeTypeCheck should be valid
AbstractMetaTypes; throw an exception if this fails.

Define a check function for Qml's VolatileBool where it was relying on
heuristics.

Task-number: PYSIDE-1660
Change-Id: I9641c63ae2942db86dbf84488429b7e25f0491a0
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>