shiboken6: Fix an impossible reference to type_new
[ChangeLog][shiboken6] It is now possible to derive classes from
enum types like `QtCore.Key` in earlier Python versions.

In SbkEnumTypeTpNew the reference to `PyType_Type.tp_new` was
made through `PyType_GetSlot` in a Limited API compatible way.

Unfortunately, `PyType_GetSlot` works for non-heaptypes in
Python 3.10, only, and `PyType_Type is not a heaptype.

This was found while creating a minimum example to prove that
the PyPy implementation of type_new is not correct.

Task-number: PYSIDE-535
Change-Id: Ibe300f4670d9db6d29fbdeb1d147e1a074ec23c6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>