shiboken: improve the handling of PyType_Type.type_new
The function type_new is needed in a few places. When the Limited
API is used, the direct access to this function is normally not possible.

We have our own work-around to overcome this limitation, but in
the case of this function, it is an advantage to have it defined
externally, since we also can return a more convenient type object.

Change-Id: I6b621051182d24e1f1469377b4570871e75dac48
Task-number: PYSIDE-535
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>