Properly specify extra include for Qt core module source file
Replace the hack adding the include to a primitive by an extra-include
element. This is actually implemented in code, but not documented.
As a drive-by fix the code to only write the comment when includes
are present.
Adapt the documentation accordingly.

Task-number: PYSIDE-1660
Change-Id: I06520f4747d02f2b3e86f90c09220d82e91f95ff
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit 418f43f7011281bd2d78f2aa43bb6b3cc666220d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>