CMake: Clean up dumpcodemodel build system code
Don't depend on apiextractor_SOURCE_DIR to get the apiextractor
include paths, in the future the tests might not be built
as part of the apiextractor project, thus the variable would be empty.
Instead use an explicit path to the directory.

Don't publicly link against apiextractor and Core, PUBLIC usage
requirements don't make sense for executables.

Pick-to: 6.2
Change-Id: I145caab0d8bd2328237ef3416b54a9eed8632f1c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>