CMake: Search for shiboken package in the PySide project earlier
When doing a pure CMake build of the PySide project (without
setup.py), configuration would fail with

 CMake Error at CMakeLists.txt:49 (message):
  Could not identify PySide6 version.  Error

because the project expects the PYTHON_EXECUTABLE cache var to be set
to a valid value. This value is usually passed by the setup.py script.

Move the find_package(Shiboken) call before the version detection
code.
This allows not passing an explicit PYTHON_EXECUTABLE cache var
because the Shiboken package will set it.

Pick-to: 6.2
Change-Id: I97d7cae07711be17c2390ac0f9db835b8d74e771
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>