CMake: pyside: Add support for cross-building
The pyside project will be cross-compiled either if
CMAKE_CROSSCOMPILING is set to TRUE (which is usually set
by a CMake toolchain file or computed by CMake itself)
or if QFP_SHIBOKEN_HOST_PATH is provided on the command line.

Various code is added to ensure the correct target Qt, Python and
Shiboken libraries are found, as well as host Qt and shiboken tools.
These are specified to the project by setting one of the following
vars:
 - QFP_QT_TARGET_PATH (for device Qt libs and includes)
 - QFP_PYTHON_TARGET_PATH (for device python libs and includes)
 - QFP_SHIBOKEN_TARGET_PATH (for device libs)
 - QFP_SHIBOKEN_HOST_PATH (for host shiboken generator)
 - QT_HOST_PATH (for host moc and friends)

When cross-compiling, pyi file generation is disabled because it's not
possible to run a target python interpeter on a host machine. It might
be possible to do that by using qemu userland emulation in the future.

Task-number: PYSIDE-802
Task-number: PYSIDE-1033
Change-Id: Ifa101e90d83397fa19132f9f0ce21e03b3523a74
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>