setup.py: Allow specifying Qt without adding it to PATH
Fix setup.py to allow specifying Qt either via --qtpaths or --qmake
options.

Previously if only one of the options were provided, the build would
fail saying that the other option was not provided.
Instead, only fail if neither options were provided.
Reconstruct the path to the missing tool based on the path of the one
that was provided.
Prefer qtpaths, because the qmake option is deprecated.

Even if both --qtpaths and --qmake options were provided to setup.py,
the build would still fail when running _get_cmake_mkspecs_variables
because it couldn't find a Qt package.
The CMake project that is configured in that function depended on
PATH being set to Qt's prefix, which CMake's find_package would then
pick up.

Explicitly pass the Qt prefix via CMAKE_PREFIX_PATH when configuring
the project.

Pick-to: 6.2
Change-Id: Ibd0cb35f8902e3d795c308c9c755a342118b0848
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>