Fix QIcon.addPixmap() to accept a PyPathLike argument
By modifying functions to accept a PyPathLike argument, we
saw the side-effect of disabling implicit conversions.

In an alternative branch, we tried to re-enable implicit
conversion by adding new functions with a PyPathLike argument.
This worked, but had drawbacks:

* the signatures become redundant, and some post-processing
  must be implemented

* the implicit conversion works fine, but only with a string
  argument. Much better would be to supply a PyPathLike

This patch leaves the modifications of function arguments
and simply adds the missing icon.addPixmap(PyPathLike).

Other implicit conversions which might be found missing
should be added the same way.

[ChangeLog][shiboken6] The implicit conversion of icon.addPixmap(str)
was replaced by an explicit version which takes PyPathLike.

Change-Id: I48a2887e28473718f027059df2aafdd516f66dc3
Fixes: PYSIDE-1669
Task-number: PYSIDE-1499
Pick-to: 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>