shiboken: fix missing cleanup in overridden virtual methods of wrappers
In an error condition, all arguments must be invalidated
in order to get a clean state.

This is currently not done when an error condition exists.

Rewrite the code to generate in a way that all
destructions are done, regardless of function outcome.

That needed some reordering of the code, in order to
move the necessary cleanup right after the Python function
call and before PyErr_Print/return.

[ChangeLog][shiboken6] Overridden virtual methods are now
always correctly finalized, regardless of errors.

Change-Id: If87cc4631f9b5731c1b1b8c7bf204e72d8744cc7
Fixes: PYSIDE-656
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 3e5bfbff99334b8c15ee705079d5a7f0cfa371d5)