shiboken6/Clang parser: Refactor include checking
In the base class BaseVisitor of the clang parser, check whether the
file has changed by comparing the CXFile. If it has changed, convert
the file name to a QString and call visitLocation().

The code is then faster since the file name checks are only performed
when the CXFile changes. The check code operating on QString becomes
simpler and can be extended more easily.

Task-number: PYSIDE-802
Task-number: PYSIDE-1660
Pick-to: 6.2
Change-Id: I1dd22ef91a4a7b96d05507b43511313507281fd6
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>