shiboken6: Implement opaque containers for getters returning a const reference
Add a bool m_const member to the container helper template that
indicates a const container. Error out of the modification functions
if that is set.  Create an additional creation function for the const
case. A const opaque containers is then of same Python type as the
non-const version, requiring no further type checks.

Pick-to: 6.2
Task-number: PYSIDE-1605
Change-Id: I45faeb0d68e6144a9dfbe25497694b8acdd98c09
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>