shiboken6: Improve fixing of enum default values of function arguments
shiboken needs to fully qualify default arguments of functions so that
they can be used outside the respective's class scope. This happens in
AbstractMetaBuilderPrivate::fixDefaultValue() and
ShibokenGenerator::guessScopeForDefaultValue(). This was confusing
particularly for enums and flags, where
AbstractMetaBuilderPrivate::fixDefaultValue() was doing a simple
expansion and further cases were handled in
ShibokenGenerator::guessScopeForDefaultFlagsValue().

To improve this and get it tested, move the enum/flags code completely
over to AbstractMetaBuilder, restructure it, rewrite the helpers
in terms of QStringView and make it a bit smarter:

- Check for hex numbers as well.
- No longer add a cast when constructing QFlags from a number(saves
  the need to observe the protected hack).
- Expand expressions "A | B" as was before and additionally
  within a type cast "Options(A | B)".
- Add a test.

Task-number: PYSIDE-1691
Change-Id: I198995508e8ed01710a44708c425c31308908445
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 5888236a23f277e1f795052d5ed93a0430d834a4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>