shiboken6: Remove ShibokenGenerator::guessScopeForDefaultValue()
Move resolving of class fields and enum values as argument default
values into AbstractMetaBuilder.

Handling of static class field constants was spread between
AbstractMetaBuilderPrivate::fixDefaultValue() and
ShibokenGenerator::guessScopeForDefaultValue().

The former was handling it for arguments of non-primitive type only
and not completely expanding namespaces. The latter was handling it
for arguments of primitive types, too, but also added some code for
non-static fields, which cannot be used as default arguments in C++.

ShibokenGenerator::guessScopeForDefaultValue() was handling enum
values for primitive and values, excluding macros by regex, but
otherwise not checking if the term is really an enum value.

Rewrite the code in AbstractMetaBuilderPrivate::fixDefaultValue()
without regexes for clarity, let it check fields and
enum values correctly via code model and fully expand namespaces.

Add tests.

Adapt the signature module to the now fully qualified signatures.

[ChangeLog][shiboken6] When qualifying function argument default
values for the generated code, shiboken no longer considers each
identifier it cannot otherwise find as an enum value and no longer
adds the class scope to it. This may require manually adding some
replace-default-expression modifications.

Task-number: PYSIDE-1691
Pick-to: 6.2
Change-Id: Id4cd2ca1f91db8c1663d7fc31e4b4ef72a5690f1
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Tismer <tismer@stackless.com>