PyPySide: fix quite a few tests using del or __del__
In PyPy, the __del__ method is only triggered
reliably by calling gc.collect()

Also, the del statement does a deletion from the namespace,
but the real deletion happens when gc.collent() is called.

This was applied to all tests which use del, regardless if
it has a visible effect on PyPy or not. It turned out to
save more 8 errors, which is great.

Task-number: PYSIDE-535
Change-Id: I6a58d90629c9eafec7307c17f021251113b3c7f2
Pick-to: 6.2
Reviewed-by: Christian Tismer <tismer@stackless.com>