Signature: re-implement error messages for generic types
Correct error messages for generic types are a
non-trivial problem that was a long-standing FIXME.

For typing.Union, typing.Iterable and typing.Sequence
this is now implemented. Other generic cases will follow.

The problem is the isinstance function which does not work
for indexed generic types. It must be implemented by hand,
broken up and handled recursively with `any` and `all`
expressions.

[ChangeLog][PySide6] Error message are more correct now when indexed
generic types are involved like Union, Sequence and Iterable.

Task-number: PYSIDE-1675
Change-Id: Idb9546bcc9dc02801c19a95f51cdbc8ca5427fbb
Pick-to: 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>