PyPySide: circumvent an lldebug regression
When building PyPy with lldebug or lldebug0 selected a
a build option, there is a regression in the signature
module that crashes every build

This crash is a confirmed PyPy bug with global
dictionaries that occurs sometime and is hard to fix
on the PyPy side.

The workaround is to skip signature initialization in
such a PyPy debug build.

Task-number: PYSIDE-535
Change-Id: I9977e5ff82759bfd882d55785528dd276012e047
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 55eebea1a48cbf5c2caef5b89226ad5c1fb18971)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>